Issue-459: Moved ICoreMapper to core module
